+++ nowcasting_dataloader/data_sources/nwp/nwp_model.py
@@ -99,9 +99,9 @@ class NWPML(DataSourceOutputML):
                 image_size_pixels,
                 image_size_pixels,
                 number_nwp_channels,
-            ),
-            x=np.sort(np.random.randn(batch_size, image_size_pixels)),
-            y=np.sort(np.random.randn(batch_size, image_size_pixels))[:, ::-1].copy()
+            ).astype(np.float32),
+            x=np.sort(np.random.randn(batch_size, image_size_pixels).astype(np.float32)),
+            y=np.sort(np.random.randn(batch_size, image_size_pixels).astype(np.float32))[:, ::-1].copy()
             # copy is needed as torch doesnt not support negative strides
             ,
             target_time=time_5,
